+++ core/src/main/java/me/wolfyscript/utilities/util/particles/animators/AnimatorCircle.java
@@ -24,6 +24,7 @@ import com.fasterxml.jackson.annotation.JsonSetter;
 import me.wolfyscript.utilities.util.math.MathUtil;
 import me.wolfyscript.utilities.util.NamespacedKey;
 import me.wolfyscript.utilities.util.particles.ParticleEffect;
+import me.wolfyscript.utilities.util.particles.shapes.ShapeCircle;
 import me.wolfyscript.utilities.util.particles.timer.Timer;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
