+++ core/src/main/java/me/wolfyscript/utilities/util/particles/animators/AnimatorShape.java
@@ -20,6 +20,8 @@ package me.wolfyscript.utilities.util.particles.animators;
 
 import me.wolfyscript.utilities.util.NamespacedKey;
 import me.wolfyscript.utilities.util.particles.ParticleEffect;
+import me.wolfyscript.utilities.util.particles.shapes.Shape;
+import me.wolfyscript.utilities.util.particles.shapes.ShapeCircle;
 import me.wolfyscript.utilities.util.particles.timer.Timer;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
