+++ core/src/main/java/me/wolfyscript/utilities/util/particles/shapes/ShapeSphere.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.wolfyscript.utilities.util.particles.animators;
+package me.wolfyscript.utilities.util.particles.shapes;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -24,27 +24,27 @@ import com.fasterxml.jackson.annotation.JsonSetter;
 import me.wolfyscript.utilities.util.NamespacedKey;
 import org.bukkit.util.Vector;
 
+import java.util.function.BiFunction;
 import java.util.function.Consumer;
-import java.util.function.Function;
 
-public class ShapeCircle extends Shape {
+public class ShapeSphere extends Shape {
 
-    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("circle");
+    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("sphere");
 
     private final double radius;
     private int resolution;
     private Direction direction;
     @JsonIgnore
-    private Function<Double, Vector> createVector;
+    private BiFunction<Double, Double, Vector> createVector;
 
     /**
      * Only used for Jackson deserialization.
      */
-    ShapeCircle() {
+    ShapeSphere() {
         this(1, 3, Direction.Y_AXIS);
     }
 
-    public ShapeCircle(double radius, int resolution, Direction direction) {
+    public ShapeSphere(double radius, int resolution, Direction direction) {
         super(KEY);
         this.radius = radius;
         this.resolution = resolution;
@@ -53,8 +53,10 @@ public class ShapeCircle extends Shape {
 
     @Override
     public void drawVectors(double time, Consumer<Vector> drawVector) {
-        for (double i = 0; i <= Math.PI * 2; i += Math.PI / resolution) {
-            drawVector.accept(createVector.apply(i));
+        for (double i = 0; i <= 2 * Math.PI; i += Math.PI / resolution) {
+            for (double j = 0; j <= Math.PI * 2; j += Math.PI / resolution) {
+                drawVector.accept(createVector.apply(j, i));
+            }
         }
     }
 
@@ -79,23 +81,26 @@ public class ShapeCircle extends Shape {
     }
 
     @JsonSetter
-    private void setDirection(Shape.Direction direction) {
+    private void setDirection(Direction direction) {
         this.direction = direction;
         createVector = switch (direction) {
-            case X_AXIS -> (t) -> {
-                double y = radius * Math.sin(t);
-                double z = radius * Math.cos(t);
-                return new Vector(0, y, z);
+            case X_AXIS -> (i, j) -> {
+                double x = radius * Math.cos(j);
+                double y = radius * Math.sin(i) * Math.sin(j);
+                double z = radius * Math.cos(i) * Math.sin(j);
+                return new Vector(x, y, z);
             };
-            case Y_AXIS -> (t) -> {
-                double x = radius * Math.cos(t);
-                double z = radius * Math.sin(t);
-                return new Vector(x, 0, z);
+            case Y_AXIS -> (i, j) -> {
+                double x = radius * Math.cos(i) * Math.sin(j);
+                double y = radius * Math.cos(j);
+                double z = radius * Math.sin(i) * Math.sin(j);
+                return new Vector(x, y, z);
             };
-            case Z_AXIS -> (t) -> {
-                double x = radius * Math.cos(t);
-                double y = radius * Math.sin(t);
-                return new Vector(x, y, 0);
+            case Z_AXIS -> (i, j) -> {
+                double x = radius * Math.cos(i) * Math.sin(j);
+                double y = radius * Math.sin(i) * Math.sin(j);
+                double z = radius * Math.cos(j);
+                return new Vector(x, y, z);
             };
         };
     }
